### 2018-09-17 #959 @whedon
[PRE REVIEW]: MTSS: A CUDA software for the analysis of Multivariate Time Series
**Submitting author:** @DavideNardone (<a href="http://orcid.org/0000-0003-0486-1791">Davide Nardone</a>)
**Repository:** <a href="https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare" target ="_blank">https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare</a>
**Version:** v1.0.0
