### 2018-10-10 #959 @DavideNardone
@danielskatz Is eventually possibile change the handling editor? The review might take more than what it's written in the `Author Guidelines`.
Thanks.

Regards,
DN
