### 2018-10-25 #1049 @jedbrown
@krischer @karlrupp :wave: Welcome! The comments from whedon above outline the review process. I'll be watching this thread if you have any questions. Thanks!

@DavideNardone It looks like there are some capitalization issues in your bibliography and perhaps you can cite a more recent CUDA since your package requires 5.0 or later.
