### 2018-10-25 #1049 @DavideNardone
@jedbrown I updated the CUDA reference in the bibliography but regarding the capitalization issue in the bibliography I generated it by using [Google Scholar Button](https://chrome.google.com/webstore/detail/google-scholar-button/ldipcbpaocekfooobnbcddclnhejkcpn?hl=en) so i dont know what to fix. Can you please be more specific ?
