### 2018-10-25 #1049 @jedbrown
@DavideNardone It is necessary to add a brace around words that should be capitalized, such as `{IEEE}`. I'll see if we can change the style file because it's a common issue and I'm not sure the lack of capitalization here is intentional.

Note that you have misspelled "Software" as "Sofwtare" in your repository name.
