### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#2 @krischer
Tag a version and create a github release for version 1.0.0
Needed for https://github.com/openjournals/joss-reviews/issues/1049
