### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#3 @krischer
Fix installation readme
The current readme clones the repository and then specifies to unzip the `master.zip` file. This is inconsistent - please fix it.

See https://github.com/openjournals/joss-reviews/issues/1049
