### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#4 @krischer
How to interpret output?
I actually don’t understand what the output of the software is. All I could do is capture the stdout but I don’t really see how to get the actual results of the classification or subspace searches. Judging from the readme this package is intended to be used as a command line utility and not a library so I assume there must be a way to get the actual results of the computation.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
