### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#5 @krischer
Non-working GPU example in readme
```bash
./mdtwObj -t SUBSEQ_SEARCH -i GPU 3 512 0 -f data/subseq_search/T_series data/subseq_search/Q_series \
     -o 3907 421 -m 1 DTW -d 0 2>&1 | tee subseq_search_GPU.txt
```

Output:
[subseq_search_GPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573782/subseq_search_GPU.txt)

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
