### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @krischer
Output different on CPU and GPU
```bash
# Classification CPU
time ./mdtwObj -t CLASSIFICATION -i CPU 3 1 \
    -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT \
    -k 10 -o 1000 152 -m 0 DTW 2>&1 | tee classification_CPU.txt

# Classification GPU
time ./mdtwObj -t CLASSIFICATION -i GPU 3 512 1 \
    -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT \
    -k 10 -o 1000 152 -m 0 DTW -d 0 2>&1 | tee classification_GPU.txt
```

These examples from the readme don't have the same results.

Results:

[classification_CPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573790/classification_CPU.txt)
[classification_GPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573791/classification_GPU.txt)

For some reason the results for the CPU and GPU versions do not match. I did not look why this is the case. If there is some random component please allow the users to set the initial state of the RNG to get reproducible results.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
