### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#7 @krischer
Performance of GPU version
I cannot reproduce the performance claims of the software.

I ran some performance test on a single Nvidia Pascal Titan X GPU and the CPU version on a single core of a Intel(R) Core(TM) i7-6850K CPU @ 3.60GHz CPU. I first ran the following two tests:

```bash
# Classification CPU
time ./mdtwObj -t CLASSIFICATION -i CPU 3 1 \
    -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT \
    -k 10 -o 1000 152 -m 0 DTW 2>&1 | tee classification_CPU.txt

# Classification GPU
time ./mdtwObj -t CLASSIFICATION -i GPU 3 512 1 \
    -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT \
    -k 10 -o 1000 152 -m 0 DTW -d 0 2>&1 | tee classification_GPU.txt
```

Results are here: [classification_CPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573805/classification_CPU.txt), [classification_GPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573806/classification_GPU.txt)

The GPU version is about a factor of 27 times faster then the single threaded CPU version which seems more realistic than the quoted three orders of magnitude. Also this is only a single core (this particular CPU has 6 cores/12 threads) so the code would have to be parallelized for a fair comparison.

Simply compiling the code with `-O3` makes the CPU version only about 6 times slower than the GPU version (results are here: ). The optimisation flags don’t make the GPU code run any faster. Given that this could be made even 6 times faster with parallelisation the claim that the GPU version is faster than the CPU version does not hold even without comparing to other implementations.

[classification_CPU_O3.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573807/classification_CPU_O3.txt), [classification_GPU_03.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2573809/classification_GPU_03.txt)

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
