### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#8 @krischer
Document the API
This package would be useful as a library but all the API functions are not documented. Please document them using doxygen or a similar tool.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
