### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#9 @krischer
Add tests
There are currently no tests or reproducible examples right now. I recommend setting up some unit and integration testing and especially compare the results against some known good result. Integration into a CI service would also be a good idea.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
