### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#10 @krischer
Add community guidelines
Please add clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
