### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#11 @krischer
Document the origin of the example data sets
It is not clear to me where the example data sets originate from or what the represent. Please document this and add DOIs where appropriate.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
