### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#12 @krischer
Some issues with the readme
* You use italic text with capitalised words for emphasis, e.g. “Time Accuracy” and “Accuracy”. This is a bit odd, can you just lower case the words if they are in the middle of a sentence?
* Sentence in the readme: “MTSS presents a GP-GPU implementation which allows it to achieve almost three order of magnitude speedup, whilst getting better Accuracy results.” This does not make sense to me, a different implementation can of course be faster but I don’t see how it should give more accurate results assuming no bugs in both of course.
* In the final “Paper” section: Typo in “is under riew” and the final sentence “so I wouldn’t recommend you to use it for any plagiarism.” is not necessary.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
