### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#13 @krischer
Add instructive examples
There are some examples in the readme but at least one of them (see answer in the functionality section) does not work. Furthermore it is not explained what these examples actually are or what the expected result should be. As mentioned before I don’t really see how to get the results.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
