### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#14 @krischer
Fix warning messages during compliation
There are quite some warning messages during the compilation stage - might be worthwhile to look into.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
