### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#15 @krischer
Add proper debug mode
“Debug mode” currently consists of a lot of commented sections `printX()` and so on. Please consider adding a proper compile option for it - this will ease the job of everyone looking at the code at a future point in time.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
