### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#16 @krischer
Consider using a proper build tool
Consider using a proper build tool like make/cmake/scons/… This would also allow for nicer error reporting for example if “-D WS” is not specified.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
