### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#17 @krischer
Code styling
Code formatting is inconsistent - please consider using a tool like `clang-format`.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
