### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#18 @krischer
Standard directory structure
Consider following the standard directory structure of `include`, `src`, …

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
