### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#19 @krischer
Command line options library
A lot of the code in `MD_DTW.cu` deals with the parsing of command line option. This is hard to read and maintain and fairly error-prone. Please consider using either a standard tool like `getopt` or another existing library (e.g. https://www.argtable.org/).

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
