### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#20 @krischer
Break apart the main function
* The main function is essentially a list of very long if/else clauses and switch statements with lots of repeated code. Please consider breaking these out in separate functions where appropriate also to reuse code more often and ease future maintenance.

* The same is true of the `MD_DTW.cu` source code. This should be more modular.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
