### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#21 @krischer
TODO comment
Line 26 and 27 of `module.cu`:

```cuda
//not working (must be changed)
__device__ float stdDev(float *data, int n,float *avg)
```

- maybe check this again?

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
