### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#22 @krischer
Remove obsolete code/commented code
There is still a lot of commented/obsolete code lines in there. Please consider cleaning them up. They will still be available in the git history should they become relevant again.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
