### 2018-11-12 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @krischer
Why is WS a compile time parameter?
I don’t see a convincing reason why `WS` must be specified at compile time. The code would be more flexible if this is a runtime variable. But maybe I am missing something. Please clarify or make it a runtime option.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
