### 2018-11-12 #1049 @krischer
# Review

This package performs dynamic time warping (DTW) in many dimensions and is implemented on the GPU. I can only recommend this package for publication after some points have been addressed and this requires a **major revision**. I encourage the author to work at these points.

For full disclaimer I note that I have never used DTW in my own research but I am familiar with similar concepts and this review here largely evaluates the software side of things which I feel qualified to do.

Major points in bullet point form (they are expanded upon later in this review and in issues in the package’s repository).

* There are no tests and benchmarks, neither against analytical toy problems nor against other software packages. A large number of libraries perform DWT so this should be done. I don’t recommend speed benchmarking against other codes as this is inherently biased and unfair but the accuracy of the results must be asserted in some way. Otherwise it is hard to trust the results of software. (https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/9)
* I don’t understand how to interpret and post-process the output of the software. (https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/4)
* I cannot reproduce the performance claims and compiling with `-O3` makes the CPU version performance competitive. (https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/7). This is a problem as speed is stated to be a major reason for the existence of this software.
* The API is not documented (https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/8)

## Comments on the checklist:

>    **Version:** Does the release version given match the GitHub release (v1.0.0)?

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/2

>    **Installation:** Does installation proceed as outlined in the documentation?

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/3

>    **Functionality:** Have the functional claims of the software been confirmed?

I don’t know to be honest. This is not my area of expertise so I don’t have a test data set to use this with. Additionally there are no unit tests with some small analytical examples that could be investigated. The readme contains some usage example and they run but I don’t know what the output should be. I recommend to add some small and easily visually verifiable examples so that this is easier to test. Also comparisons with other software packages would be good to have.

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/4
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/5
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/6

>    **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The self-proclaimed main reason of existence of the software is to be faster. The specific claim is that the GPU version is three orders of magnitudes faster. This seems a bit excessive to me as common speed ups (except for some extreme cases) when moving a well optimised code to GPUs seem to be at most an order of magnitude. In general these comparison are really hard to make and would require well-optimised versions of both codes and benchmarking in very controlled environments. The first comment in this thread is IMHO well argued: https://devtalk.nvidia.com/default/topic/945987/what-is-maximum-speed-up-that-can-be-obtained-with-gpu-/

I’d recommend removing the performance claims or backing them by some substantial benchmarks, especially also comparing against existing and optimised CPU implementations.

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/7

>    **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

No package manager but the only requirement is a working cuda compiler.

>    **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/13

>  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/8

>  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/9

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/10

> **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/11

## Comments on the readme:

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/12

## Generic notes on the code:

https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/14
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/15
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/16
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/17
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/18
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/19
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/20
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/21
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/22
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/23

## Comments on the paper:

* Please add a few more citation to the summary to back the made claims and statements. I’m sure there are review papers about DTW out there.
* According to https://joss.readthedocs.io/en/latest/review_criteria.html the “Examples” section should be removed.
* According to https://joss.readthedocs.io/en/latest/review_criteria.html the “Acknowledgements” section should be removed.
* I’m not sure if the source code link should be moved to the references. @jedbrown can you clarify?
* “Although these problems seem to be already solved in the state-of-art, nobody mentions the problem to process these tasks on large quantities of data in a reasonable amount of time.” Not my area of expertise but one of the papers you cite (Accelerating Dynamic Time Warping Subsequence Search with GPUs and FPGAs) uses GPUs and even FPGAs so I assume that people do care about this. Also “these problems seem to be already solved” is not a scientific statement. Please be clear and also provide references where appropriate.
* The R-MDTW approach is described as not being a different DTW similarity measurement but something else. This is a bit confusing to me.
* The CUDA, N. citation is not correct.
