### 2018-11-13 #1049 @DavideNardone
@krischer Thank you for reviewing the software, I really appreciate the effort you put in. Along with my work commitments, I'll try to give reasonable answers to all your questions and to fix the bugs (if it's possible) as soon as possible.
