### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @DavideNardone
As you said, a component of the software(the K-Fold Cross Validation) includes a random component which causes the GPU and CPU versions to get different results on the same initial dataset. I've identified the function who triggers that behavior and i'll fix it by adding a _shuffle_flag_ on the command line( e.g., -k 10 <0,1>) which will allow the user to get reproducible results.
