### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @krischer
Great. You don't need to close issues manually - you can just include `closes #6` in a commit message to master and the issue will be closed with a reference to the commit fixing it - that way it is easier to follow. The same works with pull requests if you prefer that workflow.
