### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#2 @DavideNardone
Yes, I've added the v1.0.0 release on the master branch.
