### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#3 @DavideNardone
I've replaced the previous two commands with the following one:
`git clone https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare.git <target>`
and therefore updated the README file.
