### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @DavideNardone
@krischer I've just read it. Thank you for the advice ;)
