### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#5 @DavideNardone
I've run the following test example and it works fine with the GPU i've used for testing the software. It might be a problem related either to the GPU device or to the CUDA version you used. I've attached here the [output](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2575973/subseq_search_GPU.txt) running of my GPU.

Anyway, I removed an extra `/` on the README file that caused the software to stop on that example test due to the wrong following path (/data/subseq_search/Q_series).
