### 2018-11-13 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#7 @DavideNardone
I understand yours concerns about the performance and I'll try to give you the best explanation for that issue. The reason why the GPU version is just roughly 6 time faster than the CPU is mainly due to the dimension of the dataset. In fact, the software is intended to work on bigger batches of data rather than just those used here as example tests. In addition, to be honest, I didn't take into account the case of comparing my software neither with a parallel version software or an optimized compiled version.

Although this comment might not be the best answer to your question, I've also found [here](https://stackoverflow.com/questions/43706755/how-can-i-get-the-nvcc-cuda-compiler-to-optimize-more) the following issue with compiling a CUDA programma with the `-03 option`  but i don't know whether this is the case or not.

Sadly, I cannot make my GPU version faster for this task since this version for the _Multi Dimensional Dynamic Time Warping_ is the best i could write. Anyhow, whoever will use the software will always benefit from it especially when the size of the dataset to handle is made of more than one thousand of time series.
