### 2018-11-15 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#7 @DavideNardone
I've uploaded two big dataset at the following [link](https://drive.google.com/open?id=1rhuKAH7XRuBaSu9LvlEJBK7tAKyqYcFJ). The information about these datasets are the summarized down here:

Name | # variables | # classes | Train set size | Test set size | Dataset size | Time series length
-- | -- | -- | -- | -- | -- | --|
**UCI_CHAR** | 9 | 6 | 7352 | 2947 | 10299 | 561
**Motor imagery in ECoG   recordings** | 64 | 2 | 278 | 100 | 378 | 3000

You can run the software on both CPU and GPU by using the following commands:

**CPU**:
`./mdtwObj -t CLASSIFICATION -i CPU 3 0 -f <path_dataset>/DATA <path_dataset>/LABEL -k 10 0 -o 10299 561 -m 0 DTW -d 0`

**GPU** :
`./mdtwObj -t CLASSIFICATION -i GPU 3 <num_threads> 0 -f <path_dataset>/DATA <path_dataset>/LABEL -k 10 0 -o 10299 561 -m 0 DTW -d 0`

I do know the _NVIDIA Visual Profiler_ tool and I will try to use it to understand whether there's a bottleneck on my code. At first glance, i might say to you that a likely bottleneck might lie in transferring the data between the host and the device.
