### 2018-11-15 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#9 @DavideNardone
I can provide some unit and integration test but i don't understand what you mean by the following sentence: _especially compare the results against some known good result?_ Could you please be more specif about it?
