### 2018-11-16 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#16 @DavideNardone
I can provide a shell script for checking nicer error reporting, like the following:

`#!/bin/bash`

`re='^[0-9]+$'`

`if test "$#" -ne 1; then`
    `echo "Error: Illegal number of parameters. \nPlease run the script with the option --help for a more clear understanding!" >&2; exit 1`
    `exit 1`
`fi`

`if ! [[ "$1" =~ $re ]] ; then`
   `echo "Error: The first argument shoul be a number" >&2; exit 1`
`fi`

`if test "$1" == "--help"; then`
	`echo 'Description help'`
	`exit 1`
`fi`

`echo 'Compiling...'`
`command1='-arch=sm_30 -D WS='$1' MD_DTW.cu module.cu  -o mdtwObj'`

`nvcc $command1`

Do you think can it be fine? I don't think is necessary to use such a tool like `cmake` for just compiling 2 files...
