### 2018-11-16 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#19 @DavideNardone
The scenario of using the `getopt` utility has already been considered but, unfortunately since the POSIX standard `getopt` does not support multiple arguments for a single flag option I had to to adopt an _ad hoc_ solution. In addition, the arguments provided by the software are not just single flag letters but in some cases array of characters. By saying that, I honestly think it's not worth to encapsulate my code into the `getopt` utility since it's application does not make any difference or makes my scenario harder to code.

However, the logic behind the code i wrote for parsing the command line options is quite similar to `getopt`. To make the code more readable and less prone to errors I agree to re-modularized it and test it better.
