### 2018-11-19 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
Compiling the sources code with the STATIC variable `WS` makes the CUDA software run faster.
Knowing a priori the length of the the time serie allows us to allocate the support matrix T directly into the registers of the CUDA's blocks where the access to it by the threads is either much faster than the shared memory or global.
Moreover, the algorithm is optimized to work only on a subset of the matrix `T`, that is `array[WS][2]`
(which is updated at each iteration) and therefore there is no problem of allocation of this structure into the registers that are much smaller than the other device memories.

![cuda-memory-model-nvidia-2007](https://user-images.githubusercontent.com/13561546/48705881-be307c00-ebfa-11e8-9ba9-ac2357556eef.png)
