### 2018-11-20 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#21 @DavideNardone
That comment is old. However, some functions in the `module.cu` was neither updated or used in the software. I haven't removed them because someone may want to use it.
