### 2018-11-20 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#15 @DavideNardone
I added a `verbose mode` for displaying some information during the running
