### 2018-11-23 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#13 @DavideNardone
All the example provided are now fully working! I have now committed a better explanation about the two tasks in the README. Furthermore, after adding the API documentation, the people interested in using the software have all the information to separately run the functions in the `module.cu`. As a plus (like stated by you on #9) i will also add some more instructive examples.
