### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @karlrupp
I've got the same question. Following your argument, every linear algebra library would only allow for squared matrices of size e.g. 128x128 because that happens to be good for caches.

The recommended approach is to make WS a runtime parameter and then dispatch between optimized/fallback implementations as needed. With a bit of metaprogramming you can auto-generate the different kernels.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
