### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#16 @karlrupp
I second this request for a build system. It is common to e.g. use something like
`./configure && make && make install` in Unix-land. Also, users on Windows benefit from build tools that auto-generate their Visual Studio project, etc.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
