### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#18 @karlrupp
The build instructions need to be updated accordingly. With source and include directories separated, a proper build tool comes in handy, cf. https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/16
