### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#14 @karlrupp
I still get the following with CUDA 8 and GCC 5.4:

```
source/module.cu(831): warning: result of call is not used

source/module.cu(839): warning: result of call is not used

source/module.cu(851): warning: result of call is not used

source/module.cu(861): warning: result of call is not used

source/module.cu(934): warning: result of call is not used

source/module.cu(940): warning: result of call is not used

source/module.cu(966): warning: result of call is not used

source/module.cu(972): warning: result of call is not used

source/module.cu(1002): warning: result of call is not used

source/module.cu(1008): warning: result of call is not used

source/module.cu(831): warning: result of call is not used

source/module.cu(839): warning: result of call is not used

source/module.cu(851): warning: result of call is not used

source/module.cu(861): warning: result of call is not used

source/module.cu(934): warning: result of call is not used

source/module.cu(940): warning: result of call is not used

source/module.cu(966): warning: result of call is not used

source/module.cu(972): warning: result of call is not used

source/module.cu(1002): warning: result of call is not used

source/module.cu(1008): warning: result of call is not used

source/module.cu: In function ‘void readFileSubSeq(char**, int*, int, float*, int, float*, int, int, int)’:
source/module.cu:831:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[0], "%f", &tmp[k]);
                                      ^
source/module.cu:839:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[1], "%f", &tmp[k]);
                                      ^
source/module.cu:851:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[0], "%f", &tmp[i]);
                                      ^
source/module.cu:861:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[1], "%f", &tmp[i]);
                                      ^
source/module.cu: In function ‘void readFile(char**, int*, int, int, float*, data, int, int*, int, int)’:
source/module.cu:934:41: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
       fscanf(inputFile[lab_ind], "%f", &label);
                                         ^
source/module.cu:940:40: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[data_ind], "%f", &tmp);
                                        ^
source/module.cu:966:35: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[k], "%f", &label);
                                   ^
source/module.cu:972:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[k], "%f", &tmp[k]);
                                      ^
source/module.cu:1002:36: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[ll], "%f", &label);
                                    ^
source/module.cu:1008:39: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
             fscanf(inputFile[ll], "%f", &tmp[j]); // fd=0 data descript
```
Part of the review at: https://github.com/openjournals/joss-reviews/issues/1049
