### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#14 @karlrupp
Update to the previous comment to better reproduce:

```
$ nvcc -Xcompiler "-O3 -Wall" -D WS=421 source/MD_DTW.cu source/module.cu -o mdtwObj -Iinclude/
nvcc warning : The 'compute_20', 'sm_20', and 'sm_21' architectures are deprecated, and may be removed in a future release (Use -Wno-deprecated-gpu-targets to suppress warning).
source/MD_DTW.cu: In function ‘int main(int, char**)’:
source/MD_DTW.cu:968:103: warning: ‘arr_num_file’ may be used uninitialized in this function [-Wmaybe-uninitialized]
source/MD_DTW.cu:41:6: note: ‘arr_num_file’ was declared here
   int *arr_num_file;
      ^
source/MD_DTW.cu:1126:11: warning: ‘distance_type’ may be used uninitialized in this function [-Wmaybe-uninitialized]
           if (strcmp(distance_type, "DTW") == 0) // DTW distance
           ^
source/MD_DTW.cu:39:31: note: ‘distance_type’ was declared here
   char *task, *compution_type, *distance_type;
                               ^
source/MD_DTW.cu:611:20: warning: ‘compution_type’ may be used uninitialized in this function [-Wmaybe-uninitialized]
       } else if (strcmp(compution_type, "GPU") == 0) {
                    ^
source/MD_DTW.cu:39:14: note: ‘compution_type’ was declared here
   char *task, *compution_type, *distance_type;
              ^
source/MD_DTW.cu:948:20: warning: ‘task’ may be used uninitialized in this function [-Wmaybe-uninitialized]
   } else if (strcmp(task, "SUBSEQ_SEARCH") == 0) {
                    ^
source/MD_DTW.cu:39:7: note: ‘task’ was declared here
   char *task, *compution_type, *distance_type;
       ^
source/MD_DTW.cu:1025:110: warning: ‘q_size’ may be used uninitialized in this function [-Wmaybe-uninitialized]
source/MD_DTW.cu:13:13: note: ‘q_size’ was declared here
   int t_size, q_size;
             ^
source/MD_DTW.cu:989:76: warning: ‘t_size’ may be used uninitialized in this function [-Wmaybe-uninitialized]
               dist = short_md_ed_c(&t_series[i], q_series, window_size, n_feat,
                                                                            ^
source/MD_DTW.cu:13:5: note: ‘t_size’ was declared here
   int t_size, q_size;
     ^
source/module.cu(831): warning: result of call is not used

source/module.cu(839): warning: result of call is not used

source/module.cu(851): warning: result of call is not used

source/module.cu(861): warning: result of call is not used

source/module.cu(934): warning: result of call is not used

source/module.cu(940): warning: result of call is not used

source/module.cu(966): warning: result of call is not used

source/module.cu(972): warning: result of call is not used

source/module.cu(1002): warning: result of call is not used

source/module.cu(1008): warning: result of call is not used

source/module.cu(831): warning: result of call is not used

source/module.cu(839): warning: result of call is not used

source/module.cu(851): warning: result of call is not used

source/module.cu(861): warning: result of call is not used

source/module.cu(934): warning: result of call is not used

source/module.cu(940): warning: result of call is not used

source/module.cu(966): warning: result of call is not used

source/module.cu(972): warning: result of call is not used

source/module.cu(1002): warning: result of call is not used

source/module.cu(1008): warning: result of call is not used

source/module.cu: In function ‘void readFileSubSeq(char**, int*, int, float*, int, float*, int, int, int)’:
source/module.cu:831:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[0], "%f", &tmp[k]);
                                      ^
source/module.cu:839:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[1], "%f", &tmp[k]);
                                      ^
source/module.cu:851:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[0], "%f", &tmp[i]);
                                      ^
source/module.cu:861:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[1], "%f", &tmp[i]);
                                      ^
source/module.cu: In function ‘void readFile(char**, int*, int, int, float*, data, int, int*, int, int)’:
source/module.cu:934:41: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
       fscanf(inputFile[lab_ind], "%f", &label);
                                         ^
source/module.cu:940:40: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[data_ind], "%f", &tmp);
                                        ^
source/module.cu:966:35: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
         fscanf(inputFile[k], "%f", &label);
                                   ^
source/module.cu:972:38: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[k], "%f", &tmp[k]);
                                      ^
source/module.cu:1002:36: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
           fscanf(inputFile[ll], "%f", &label);
                                    ^
source/module.cu:1008:39: warning: ignoring return value of ‘int fscanf(FILE*, const char*, ...)’, declared with attribute warn_unused_result [-Wunused-result]
             fscanf(inputFile[ll], "%f", &tmp[j]); // fd=0 data descript
                                       ^
source/module.cu: In function ‘float short_dtw_c(float*, float*, int, int)’:
source/module.cu:1483:11: warning: ‘g’ may be used uninitialized in this function [-Wmaybe-uninitialized]
   int k, l, g;
           ^
source/module.cu: In function ‘float short_md_dtw_c(float*, float*, int, int, int, int)’:
source/module.cu:1581:11: warning: ‘g’ may be used uninitialized in this function [-Wmaybe-uninitialized]
   int k, l, g;
           ^
```
