### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#26 @karlrupp
Document CUDA arch requirement 2.0
Since printf is used inside CUDA kernels, `arch=sm_20` is required for NVCC. Since the default arch for CUDA 5 and 6 is below `sm_20`, this requirement should be listed in the documentation.
