### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @karlrupp
I encounter similar reproducibility issues:

```
$> ./mdtwObj -t CLASSIFICATION -i CPU 3 1 -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT -k 10 0 -o 1000 152 -m 0 DTW -v 0

The number of iteration is greater than testSize! Verbose mode will be suppressed for this run
Reading data...
Dataset size: [1000,152,3]

Classification w/ DEPENDENT-DTW using CPU

Regular Accuracy mean is 53.967915
The Error rate mean of NR is 0.460321

$> ./mdtwObj -t CLASSIFICATION -i GPU 3 512 1 -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT -k 10 0 -o 1000 152 -m 0 DTW -d 0 -v 0

The number of iteration is greater than testSize! Verbose mode will be suppressed for this run
Reading data...
Dataset size: [1000,152,3]

Classification w/ DEPENDENT-DTW using GPU

Regular Accuracy mean is 4.352184
The Error rate mean of NR is 0.956478
```
