### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#7 @karlrupp
I encounter similar results (and thus conclusions) as @krischer . In my case, the example referenced in the first comment runs in 120 seconds on the CPU (Xeon E3-1240 v3) and 9 seconds on the GPU (GTX 760).

Keeping in mind that the CPU version
  - only uses a single core
  - is not as highly tuned as the GPU kernels (just looking at `short_dtw_c()` shows inefficient multi-dimensional arrays with `malloc`s and `free`s that could definitely be optimized further)
  - has no fundamental hardware characteristic that warrants a GPU speedup of more than a factor of ~10

the claims of three orders of magnitude performance gains in the paper cannot be supported.

Part of a review at: https://github.com/openjournals/joss-reviews/issues/1049
