### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#27 @karlrupp
Also the CPU path is not valgrind clean:

```
$> valgrind ./mdtwObj -t CLASSIFICATION -i CPU 3 1 -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT -k 10 0 -o 1000 152 -m 0 DTW -v 0
==28320== Memcheck, a memory error detector
==28320== Copyright (C) 2002-2015, and GNU GPL'd, by Julian Seward et al.
==28320== Using Valgrind-3.11.0 and LibVEX; rerun with -h for copyright info
==28320== Command: ./mdtwObj -t CLASSIFICATION -i CPU 3 1 -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT -k 10 0 -o 1000 152 -m 0 DTW -v 0
==28320==

The number of iteration is greater than testSize! Verbose mode will be suppressed for this run
Reading data...
Dataset size: [1000,152,3]

Classification w/ DEPENDENT-DTW using CPU

==28320== Invalid read of size 4
==28320==    at 0x409BA6: accumarray (module.cu:1253)
==28320==    by 0x409EA9: crossvalind_Kfold (module.cu:1343)
==28320==    by 0x404348: main (MD_DTW.cu:364)
==28320==  Address 0x60ccf90 is 0 bytes after a block of size 4,000 alloc'd
==28320==    at 0x4C2DB8F: malloc (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==28320==    by 0x409E14: crossvalind_Kfold (module.cu:1332)
==28320==    by 0x404348: main (MD_DTW.cu:364)
==28320==
...
```
