### 2018-11-27 #1049 @karlrupp
# Review of MTSS

MTSS provides GPU-enabled implementations for the analysis of multivariate time series. As such, the software primarily aims to provide a *fast implementation* of an established family of algorithms, rather than providing implementations of fundamentally novel algorithmic approaches. Thus, in the following I focus on the performance aspect.

## Comments on Checklist Items
  * *Version*: I did test v1.0.0 that is referenced here, but I notice that the release tag is younger than the review by @krischer . To maximize utility of my review, I thus used the latest master branch at c1b754a324be33d426829be3185558c4f9fb8a9b . Several of the issues raised in the previous review have been fixed already.
  * *Installation*: See next section
  * *Functionality*: See next section
  * *Performance*: See next section
  * *Automated tests*: Missing, see https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/9
  * *References*: No DOIs provided. Created pull request for capitalization issues raised by @jedbrown  earlier: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/pull/24

## Downloading, Compiling the Code, and Running the Examples

The repository is sufficiently light-weight in order to be cloned quickly. After cloning, the lack of a build system is apparent: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/16 Particularly now as the sources are split up into `include` and `source` folders (see https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/18), extra parameters are required to `nvcc`.

The compilation instructions in the README as well as in the paper still need to be updated accordingly. In particular, the source file extensions need to be changed back to `cu`:
https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/pull/25

Passing a worksize parameter at compile time exposes an implementation detail to the user: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/23
I recommend runtime dispatches inside MTSS. Moreover, the documentation should also mention that CUDA arch 2.0 or above is required: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/26

While the default compilation command succeeds, merely passing an optimization flag to the host compiler triggers a set of warnings: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/14

Once the compilation stage is completed, running the examples works as documented. On closer inspection, however, there are two isses:
  * *Correctness of implementation*: Running the CUDA code through `cuda-memcheck` exposes invalid memory accesses: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/27 This might also explain why I encountered different results for the CPU and GPU versions even if randomization is turned off: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/6 Valgrind also reports issues in the CPU path.
  * *Performance*: The three orders of magnitude performance gains by using GPUs as claimed in the paper cannot be reproduced at all: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/7 Instead, a quick look at the CPU implementation showed a rather inefficient use of multidimensional arrays, so the CPU performance of MTSS is certainly not a good baseline.

The API documentation is appropriate, particularly when considering that MTSS is meant to be used as a standalone software, not as a library. I did notice, however, that many functions are referred to as `callback`s, even though they are not used as callback routines in the established meaning.

I strongly recommend to add an extensive test suite: https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/issues/9
It might have caught e.g. memory access violations in an earlier and allows others to better verify the correctness of the results.

## Comments on the paper

The paper is fairly well written and explains the needs MTSS addresses. If possible, the paper would benefit from a figure (e.g. the one referenced in the README.md).

The `Examples` section would benefit from a bit of extra prose explaining the compilation and providing a bit of context to the examples.
