### 2018-11-27 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#18 @DavideNardone
Bug fixed
