### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#14 @DavideNardone
Fixed
