### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @DavideNardone
Well I do not (see attached files). You maybe compiled the sources with `WS` parameter different than 152. Anyway I will add a check on both parameters to make sure they are identical.

[repoducibility_CPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2623643/repoducibility_CPU.txt)
[repoducibility_GPU.txt](https://github.com/DavideNardone/MTSS-Multivariate-Time-Series-Software/files/2623644/repoducibility_GPU.txt)
