### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#6 @karlrupp
Well, should a user really recompile the software for every example?

(In analogy, should I recompile my browser every time I want to go to a new webpage?)
