### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
I am sorry but the software has been thought and optimized from the beginning to work with the `WS` static variable. All kernels take into account the using of that variable, therefore changing it would mean changing all the logic implemented into the software.
