### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @karlrupp
You can pass WS as a template argument and do the dispatch outside the kernels :-)
