### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
As far as I know, a static variable cannot be compiled at runtime. The following data_structure `float array[WS][2]` relies on `WS` variable to make the access to memory faster. If there's a way to bypass it I don't have any clue about it.
