### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @karlrupp
For example:
```
switch (runtime_ws) {
case 8:  optimized_kernel<8>(args);
case 16:  optimized_kernel<16>(args);
case 32:  optimized_kernel<32>(args);
case 64:  optimized_kernel<64>(args);
case 128:  optimized_kernel<128>(args);
default: fallback_kernel<96>(args);
}
```
where the template parameter is the 'static' WS (this snippet is of course just a first rough outline).

There's also the option to just-in-time compile CUDA kernels in CUDA 8 (I think) and above.
