### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
By trying this approach i get (obviously) this error: `source/module.cu(312): error: expression must have a constant value`. I don't think it's possible to do what you suggested....
