### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @karlrupp
What is the kernel function declaration that you've tried?
