### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
Sorry my bad! It compile as long as the the kernel function definition is included in the header file... For example, given the following files:

**test2.h**
```
#ifndef HEADER_H
#define HEADER_H

#ifdef __cplusplus
extern "C" {
#endif

template<int >
__global__ void kernel();

#ifdef __cplusplus
}
#endif

#endif
```
**test2.cu**
```
#include "test2.h"

using namespace std;

template<int action>
__global__ void kernel()
{
	// printf("%d\n",action);
    for(int i=0; i < action; ++i) {}

	float array[action][2];

}
```

**test.cu**
```
#include <stdlib.h>
#include <stdio.h>
#include <sys/time.h>
#include <cuda.h>
#include <unistd.h>

#include "test2.h"

int main(int argc, char **argv) {

    float grid_size;
    dim3 grid;
    dim3 threads;

    grid_size = ceil((float)100 / 2);

    grid.x = grid_size;
    grid.y = 1;
    threads.x = 10;
    threads.y = 1;

	int ws_runtime = atoi(argv[1]);
	printf("%d\n", ws_runtime);

	kernel<2><<<grid, threads>>>();
}
```

When i run `nvcc test.cu test3.cu -o test` i get the following error:

_test2.h(20): error: this declaration may not have extern "C" linkage_

Plus, I still cant find _just-in-time_ compile CUDA kernels but i suppose it will allow the following instruction `optimized_kernel<ws_runtime>(args);` to be compiled at run-time.
 In case the latter couldn't be possible i think it's craziness to hard-code all the `WS` possible parameters which theoretically could be infinite.
