### 2018-11-28 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#27 @DavideNardone
> cuda-memcheck ./mdtwObj -t CLASSIFICATION -i GPU 3 512 1 -f data/classification/rm_1/X_MAT data/classification/rm_1/Y_MAT data/classification/rm_1/Z_MAT -k 10 0 -o 1000 152 -m 0 DTW -d 0 -v 0

When i run the software i got a different error, that was due to the _first_ invalid parameter for the cudaMemset function.

```
========= Program hit cudaErrorInvalidValue (error 11) due to "invalid argument" on CUDA API call to cudaMemset.
=========     Saved host backtrace up to driver entry point at error
=========     Host Frame:/usr/lib/x86_64-linux-gnu/libcuda.so.1 [0x330453]
=========     Host Frame:./mdtwObj [0x3ed3c]
=========     Host Frame:./mdtwObj [0x4470]
=========     Host Frame:/lib/x86_64-linux-gnu/libc.so.6 (__libc_start_main + 0xf5) [0x21f45]
=========     Host Frame:./mdtwObj [0x6f25]
```
I fixed it by replacing the callback`cudaMemset(&d_test, 0, n_feat * window_size * sizeof(float))` with the following one: `cudaMemset(d_test, 0, n_feat * window_size * sizeof(float))`

By the way, I suppose the _out-of-bounds threads_ error you get by running the `cuda-memcheck` function is due to the limitation of the number of threads (< 1024) on your GPU. If so, i will provide a fix asap.
Anyway, could please provide me the `Maximum number of threads per block` of your GPU?
