### 2018-11-29 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @karlrupp
Template functions don't have external linkage. You'd have to explicitly instantiate the respective instantiations in the test2.cu file. This is not CUDA-specific; you'd have the same problem with normal C++ sources.

A simple fix is to move the kernel launch into the same file as the kernel.

You don't need to hard-code all possible WS-parameters, but you will need a reasonable strategy for selecting suitable WS sizes. For example, you can always choose the next power of 2 (maybe plus or minus 1 to avoid bank conflicts) for WS. This way all choices for WS from 65...127 map to WS=128 in the "infinite" range of parameters for WS becomes all of a sudden very small.
