### 2018-11-29 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#23 @DavideNardone
If i got you right, you're suggesting me to map range of `WS` parameters into different template kernels.
Let's hypotize I run the software with `ws_runtime=520` which would then call the `optimized_kernel<1024>(args)`. By doing this, I would allocate the data_structure `float array[ws][2]` with `ws=1024` which i think would result in a wasting `2x504` memory locations.

Do you think is worth wasting memory instead of including a static variable?
