### 2018-12-21 #1049 @DavideNardone
Hello everybody, I finally made most of the changes you've requested. Some of them remain still unsolved but the explanation is contained in its section. I've also updated the paper contents. There may be still something issues to solve it. Please let's clarify them together so that i can provide a better solution.

Thanks.

Cheers,

DN
