### 2018-12-21 #1049 @DavideNardone
> @DavideNardone - the paper won't compile as it's entirely missing the YAML header at the top of the `paper.md`. See [this example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for what's missing.

Sorry my bad!
