### 2019-01-15 #1049 @DavideNardone
Hello everybody, I finally made most of the changes you've requested. There may be still something issues to solve it but please let's clarify them together so that i can provide a better solution.
I've also tried compiling the new version of the paper but i cannot do it due to the error i got above. Can any of you give a look into this error since it does not seem so clear to me.

Thank you.

Cheers,

DN
