### 2019-01-15 #1049 @jedbrown
What error? The proof above seems well-formed. Many bib items need DOI and conference names need case protection `{ACM SIGKDD}` ...  Also, the source repository is in the margin so no need to link it manually in the text.

You should also compare to related work; what does MTSS provide to the research community that isn't available through existing packages? (It doesn't need to be unique capability, it could be better performance, a better license, improved extensibility with evidence of that, etc.)
