### 2019-01-15 #1049 @DavideNardone
There was no error, i just copied and pasted the old message wrong. Regarding the comparing with other tools, i haven't found any Multidimensional Time Series source code available yet to compare against.  Although i understand your concern about it, there's evidence that the GPU version perform faster than the CPU ones, therefore i won't be surprised whether other CPUs methods will be slower. Anyhow, what kind of comparing do you want me to perform? Can you please be more specific?

Furthermore, I don't understand what do you mean by: _"what does MTSS provide to the research community that isn't available through existing packages?"._ It's been initially stated that the software was a command line tool and not a library. If this was a prior requirement, why was this not initially specified as a fundamental requirement for the software to be published?
