### 2019-01-15 #1049 @jedbrown
GPU implementations being faster than CPU implementations is by no means automatic. If you assert that it is faster, then you'd have to provide evidence of that. If there is no other software that does what MTSS does, then you wouldn't make a claim of better performance except by a performance model (e.g., if you achieve a fraction of memory bandwidth or floating point peak). You can still mention significant software that is nearby and identify how MTSS is unique.

There is no requirement that JOSS submissions be libraries.
