### 2019-01-15 #1049 @DavideNardone
Can you please explicitly tell me where I claimed my software achieve better performance than others? Do you refer to this sentence: "This problem has been addressed in [@sanguansat2012multiple] which allows the alignment of multidimensional time signals by simply extending the concept behind the DTW but showing low-speed performance". If so, I will try to find Multivariate Time Series software with which I can compare with.
