### 2019-01-15 #1049 @jedbrown
I was responding to your comment above speculating that a GPU implementation would outperform a hypothetical CPU implementation. As for the paper, it should put its contribution in context. If there is no other software that does what MTSS does, you can state that and reviewers can evaluate whether that is true. If there is other software, you should explain the context in a way that would help a researcher decide what software is most appropriate for their needs.
