### 2019-01-21 #1049 @DavideNardone
Well, the fact a GPU implementation could be not faster than a CPU one can be true depending on the kind of task. I honestly think this kind of task is a good fit for the GPU when the dataset is made of many sample and variables. In the latter case, MTSS can drastically improve the performance of the comparisons since our I-MDTW version can split each DTW variable computation over different threads. As far as i know MTSS is the only GPU software able to compute a Multivariate Dynamic Time Warping.

However, I've been looking for some other software to compare against to but since now I've found a M-DTW CPU based version (also cited in the work) that I could use to make some comparisons.
What do you think about it?
