### 2019-01-22 #1049 @jedbrown
Yes, if performance is an explicit claim, then comparison to an existing state of the art is warranted. Reviewers may chime now or later in if they are aware of specific packages that would be good to compare with.
