### 2019-01-25 #1049 @jedbrown
@DavideNardone Your submission was in rough shape and required major revisions, so you don't get to complain about time since initial submission. I hope our reviewers get a chance to review the latest updates in the next couple weeks. Did you add performance results? (You might need to `@whedon generate pdf` if you revised the paper after the discussion above.)
