### 2019-01-25 DavideNardone/MTSS-Multivariate-Time-Series-Sofwtare#9 @jedbrown
The reference is to *verification*: establishing that the code computes the mathematical quantity it claims to. One could choose special inputs such that the outputs are known without running your code. Alternatively, you could compare to outputs produced by some trusted community code.
