### 2019-02-18 #1049 @DavideNardone
Hello everybody,

I've got some of the results but i haven't had time to update the paper yet. I will asap.

Here a preview for the first task.
