### 2019-02-25 #1049 @DavideNardone
Hello everybody, since i couldn't find more than one or two existing state of the art Multivariate Dynamic Time Warping software to which compare my software I haven't completed the tests yet. I've been thinking whether to compare my software against CPU-Univariate Dynamic Time Warping. Whether just one/two methods will not be fine for you I think that I will need more time to complete the tests. Said that, let me know whether just using two existing state of the art methods is sufficient.

Kind Regards,
DN
