### 2019-02-26 #1049 @jedbrown
It isn't so much about the number of packages that you compare to, but about whether the comparisons are representative and whether the packages you compare to are the most relevant to a researcher trying to choose appropriate software. If your software is also competitive for univariate DTW, then that might also be a draw for attracting users.
