### 2019-04-05 #1049 @jedbrown
@DavideNardone Please check you citation format. Separate multiple citations with semicolon and choose whether they are in-text.  https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
Please also check capitalization of your references (some need to be protected). Please also consider changing your long first paragraph with numbered parts to use a bulleted or numbered list formatting.
