### 2019-04-05 #1049 @DavideNardone
@jedbrown how can i manipulate images (resize, stretch, etcc ) ?

Thank you
