### 2019-04-05 #1049 @DavideNardone
@jedbrown how do I manipulate images (resize, stretch, etcc ) ?

Thank you
