### 2019-04-05 #1049 @jedbrown
Your figures would be much better as vector graphics, e.g., PDF includes. You can set width by adding `![caption](file.png){width=4in}`.

Please check in-text and capitalization for your references.
