### 2019-04-05 #1049 @DavideNardone
> Your figures would be much better as vector graphics, e.g., PDF includes. You can set width by adding `![caption](file.png){width=4in}`.
>
This command does not work for me..
