### 2019-04-05 #1049 @DavideNardone
It doesn't work on Github preview. Anyway I am sick of your fussiness about everything and I want you to remove my software from your """journal""".
