### 2019-04-05 #1049 @jedbrown
1. Your repository does not contain a commit adding `{width=...}` to a figure. I don't know how you concluded that the syntax did not work without pushing it to the repo.
2. Formatting/style standards are a normal part of research publishing. JOSS does not have staff to do post-acceptance typesetting, thus editors/reviewers will point out those issues.
3. JOSS generally does not "reject" papers after starting a review. Instead, it strives for a constructive review to raise the submission up to the standards for publication. That takes quite a bit of effort in some cases. Of course you may withdraw your submission. Is that what you are requesting?
