### 2019-04-08 #1049 @labarba
👋 @DavideNardone — I am the Associate Editor in Chief on rotation this week. I want to tell you that the volunteer team of editors and reviewers for JOSS dutifully works to improve the quality of the software submitted, and the communication about the software (via the paper and the documentation) to potential users. Some of the process is automated (like checking all DOIs of the references properly resolve, using the `@whedon check references` command), but most of it is careful manual labor—like checking capitalization in titles of cited works, layout improvements, and copy editing.

This is meant to be a constructive process of collaboration with authors. Our editors and reviewers give their time to this enterprise because we all believe that this is a valuable model for publishing.

Do you wish to continue this process until publication?
