### 2019-04-09 #1049 @DavideNardone
Hello there,

I think the work I've done is more than enough for this journal. If you don't think so then I will not agree on doing any other work and therefore it's better to proceed with a withdraw.

Sincerely,
DN
