### 2019-04-09 #1049 @labarba
@DavideNardone, it sounds like you feel like you are working _for_ the journal, when in fact it's the journal (and its volunteers) that is working for you. It's a bit sad, but we also care about looking after our collegial community, and your demand to accept your terms is incompatible with our ethos.
